ports passing first unit tests [skip ci]
